+++ packages/react-native/ios/swiftpm/empty.swift
@@ -1,0 +1,2 @@
+// Deliberately left empty: We need a SwiftPM target with sources, but the only responsibility of
+// the build is to link the core extension.