+++ packages/web/src/db/sync/SharedWebStreamingSyncImplementation.ts
@@ -59,6 +59,10 @@ class SharedSyncClientProvider extends AbstractSharedSyncClientProvider {
     await this.options.uploadCrud();
   }
 
+  override async postCheckpointRequest(clientId: string, requestId: string): Promise<string | null> {
+    return await this.options.postCheckpointRequest(clientId, requestId);
+  }
+
   get logger() {
     return this.options.logger;
   }
@@ -187,11 +191,6 @@ export class SharedWebStreamingSyncImplementation extends WebStreamingSyncImplem
     return this.syncManager.disconnect();
   }
 
-  async getWriteCheckpoint(): Promise<string> {
-    await this.waitForReady();
-    return this.syncManager.getWriteCheckpoint();
-  }
-
   async dispose(): Promise<void> {
     await this.waitForReady();
 
@@ -226,6 +225,10 @@ export class SharedWebStreamingSyncImplementation extends WebStreamingSyncImplem
 
   async waitForReady() {
     return this.isInitialized;
+  }
+
+  requestCheckpoint(): Promise<bigint> {
+    return this.syncManager.requestCheckpoint();
   }
 
   updateSubscriptions(subscriptions: SubscribedStream[]): void {